mara-hcp/simulators/gpu-sim/requirements.txt
kafka-python==2.0.2
asyncio==3.4.3
dataclasses-json==0.6.3
psutil==5.9.6
prometheus-client==0.19.0

